Created first Branch
Added profile image from github as none was provided.
Changed some content and left the markup and style the same as my Sr-1 branch. Much of the content is not included as this was more of a html part compared to Resume, will need to discuss Resume content with Kanti once all branches are created and reviewed.